tools - change updateArray
